+++ src/Controller.php
@@ -775,7 +775,10 @@ class Controller extends Speaker
         return $this;
     }
 
+
     /**
+     * Get the network instance used by this controller.
+     *
      * @return Network
      */
     public function getNetwork()